refactor(twitter): extract RequestQueue class for improved request handling
- Moved the `RequestQueue` implementation from `base.ts` to a new `RequestQueue.ts` file, enhancing modularity and separation of concerns.
- The `RequestQueue` class now manages request queuing and processing, including exponential backoff and random delays, improving the overall structure of the codebase.
