+++ packages/client-twitter/src/RequestQueue.ts
@@ -1,0 +1,55 @@
+import { elizaLogger } from "@elizaos/core";
+
+const MS_IN_SEC = 1000;
+const DELAY_MS = 2000;
+const DELAY_RANGE_MS = 1500;
+
+export class RequestQueue {
+    private queue: (() => Promise<any>)[] = [];
+    private processing: boolean = false;
+
+    async add<T>(request: () => Promise<T>): Promise<T> {
+        return new Promise((resolve, reject) => {
+            this.queue.push(async () => {
+                try {
+                    const result = await request();
+                    resolve(result);
+                } catch (error) {
+                    reject(error);
+                }
+            });
+            this.processQueue();
+        });
+    }
+
+    private async processQueue(): Promise<void> {
+        if (this.processing || this.queue.length === 0) {
+            return;
+        }
+        this.processing = true;
+
+        while (this.queue.length > 0) {
+            const request = this.queue.shift()!;
+            try {
+                await request();
+            } catch (error) {
+                elizaLogger.error("Error processing request:", error);
+                this.queue.unshift(request);
+                await this.exponentialBackoff(this.queue.length);
+            }
+            await this.randomDelay();
+        }
+
+        this.processing = false;
+    }
+
+    private async exponentialBackoff(retryCount: number): Promise<void> {
+        const delay = Math.pow(2, retryCount) * MS_IN_SEC;
+        await new Promise((resolve) => setTimeout(resolve, delay));
+    }
+
+    private async randomDelay(): Promise<void> {
+        const delay = Math.floor(Math.random() * DELAY_MS) + DELAY_RANGE_MS;
+        await new Promise((resolve) => setTimeout(resolve, delay));
+    }
+}
