+++ packages/client-twitter/src/base.ts
@@ -18,6 +18,7 @@ import {
 import { EventEmitter } from "events";
 import { TwitterConfig } from "./environment.ts";
 import { TwitterAuthManager } from "./TwitterAuthManager.ts";
+import { RequestQueue } from "./RequestQueue.ts";
 
 type TwitterProfile = {
     id: string;
@@ -26,56 +27,6 @@ type TwitterProfile = {
     bio: string;
     nicknames: string[];
 };
-
-class RequestQueue {
-    private queue: (() => Promise<any>)[] = [];
-    private processing: boolean = false;
-
-    async add<T>(request: () => Promise<T>): Promise<T> {
-        return new Promise((resolve, reject) => {
-            this.queue.push(async () => {
-                try {
-                    const result = await request();
-                    resolve(result);
-                } catch (error) {
-                    reject(error);
-                }
-            });
-            this.processQueue();
-        });
-    }
-
-    private async processQueue(): Promise<void> {
-        if (this.processing || this.queue.length === 0) {
-            return;
-        }
-        this.processing = true;
-
-        while (this.queue.length > 0) {
-            const request = this.queue.shift()!;
-            try {
-                await request();
-            } catch (error) {
-                elizaLogger.error("Error processing request:", error);
-                this.queue.unshift(request);
-                await this.exponentialBackoff(this.queue.length);
-            }
-            await this.randomDelay();
-        }
-
-        this.processing = false;
-    }
-
-    private async exponentialBackoff(retryCount: number): Promise<void> {
-        const delay = Math.pow(2, retryCount) * 1000;
-        await new Promise((resolve) => setTimeout(resolve, delay));
-    }
-
-    private async randomDelay(): Promise<void> {
-        const delay = Math.floor(Math.random() * 2000) + 1500;
-        await new Promise((resolve) => setTimeout(resolve, delay));
-    }
-}
 
 export class ClientBase extends EventEmitter {
     static _twitterClients: { [accountIdentifier: string]: Scraper } = {};
